Change spelling
Rename d06e to d09e.

We cannot do this for the stronghold puzzle because then the shortest
solution is 14 characters, which might be too long for brute force
solutions to work within a reasonable amount of time. The SMT
solver-based solution still runs in less than a second, but we cannot
rely on people using that kind of solution.